+++ interface/mmal/core/mmal_port_clock.c
@@ -102,8 +102,8 @@ static MMAL_STATUS_T mmal_port_clock_process_buffer(MMAL_PORT_T *port, MMAL_BUFF
 
    if (buffer->length != sizeof(MMAL_CLOCK_EVENT_T))
    {
-      LOG_ERROR("invalid buffer length %d expected %ld",
-                buffer->length, sizeof(MMAL_CLOCK_EVENT_T));
+      LOG_ERROR("invalid buffer length %d expected %d",
+                buffer->length, (int)sizeof(MMAL_CLOCK_EVENT_T));
       return MMAL_EINVAL;
    }
